新增operate-webapp模块

+++ camunda/operate/webapp/src/main/java/io/camunda/operate/webapp/management/UsageMetricsService.java
@@ -1,3 +1,15 @@
+/*
+ * Decompiled with CFR 0.152.
+ * 
+ * Could not load the following classes:
+ *  io.camunda.operate.es.contract.MetricContract$Reader
+ *  io.camunda.operate.webapp.management.dto.UsageMetricDTO
+ *  io.camunda.operate.webapp.management.dto.UsageMetricQueryDTO
+ *  org.springframework.beans.factory.annotation.Autowired
+ *  org.springframework.boot.actuate.endpoint.web.annotation.RestControllerEndpoint
+ *  org.springframework.stereotype.Component
+ *  org.springframework.web.bind.annotation.GetMapping
+ */
 package io.camunda.operate.webapp.management;
 
 import io.camunda.operate.es.contract.MetricContract;
@@ -9,28 +21,20 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Component
-@RestControllerEndpoint(
-   id = "usage-metrics"
-)
+@RestControllerEndpoint(id="usage-metrics")
 public class UsageMetricsService {
-   @Autowired
-   private MetricContract.Reader reader;
+    @Autowired
+    private MetricContract.Reader reader;
 
-   @GetMapping(
-      value = {"/process-instances"},
-      produces = {"application/json"}
-   )
-   public UsageMetricDTO retrieveProcessInstanceCount(UsageMetricQueryDTO query) {
-      Long total = this.reader.retrieveProcessInstanceCount(query.getStartTime(), query.getEndTime());
-      return (new UsageMetricDTO()).setTotal(total);
-   }
+    @GetMapping(value={"/process-instances"}, produces={"application/json"})
+    public UsageMetricDTO retrieveProcessInstanceCount(UsageMetricQueryDTO query) {
+        Long total = this.reader.retrieveProcessInstanceCount(query.getStartTime(), query.getEndTime());
+        return new UsageMetricDTO().setTotal(total.longValue());
+    }
 
-   @GetMapping(
-      value = {"/decision-instances"},
-      produces = {"application/json"}
-   )
-   public UsageMetricDTO retrieveDecisionInstancesCount(UsageMetricQueryDTO query) {
-      Long total = this.reader.retrieveDecisionInstanceCount(query.getStartTime(), query.getEndTime());
-      return (new UsageMetricDTO()).setTotal(total);
-   }
+    @GetMapping(value={"/decision-instances"}, produces={"application/json"})
+    public UsageMetricDTO retrieveDecisionInstancesCount(UsageMetricQueryDTO query) {
+        Long total = this.reader.retrieveDecisionInstanceCount(query.getStartTime(), query.getEndTime());
+        return new UsageMetricDTO().setTotal(total.longValue());
+    }
 }
